refactor app
- Extracted Retrofit setup to data layer via RetrofitInstance
- Created ApiService, PhotoRepository, and connected them to ViewModel
- Implemented ViewModelFactory for proper ViewModel instantiation with dependencies
- Separated UI concerns into composable files: ImageCard, LoadingScreen, ErrorScreen
- Improved code structure, readability, and testability
