+++ app/src/main/java/mrcoder/instaexplore/jetpackcompose/UiState.kt
@@ -1,0 +1,7 @@
+package mrcoder.instaexplore.jetpackcompose
+
+sealed class UiState<out T> {
+    data object Loading : UiState<Nothing>()
+    data class Success<out T>(val data: T) : UiState<T>()
+    data class Error(val message: String) : UiState<Nothing>()
+}
