+++ app/src/main/java/mrcoder/instaexplore/jetpackcompose/data/AppContainer.kt
@@ -1,0 +1,5 @@
+package mrcoder.instaexplore.jetpackcompose.data
+
+object AppContainer {
+    val photoRepository = PhotoRepository()
+}
