+++ app/src/main/java/mrcoder/instaexplore/jetpackcompose/data/PhotoRepository.kt
@@ -1,0 +1,9 @@
+package mrcoder.instaexplore.jetpackcompose.data
+
+import mrcoder.instaexplore.jetpackcompose.model.Photo
+
+class PhotoRepository() {
+    suspend fun fetchPhotos(offset: Int, limit: Int): List<Photo> {
+        return RetrofitInstance.api.getPhotos(offset, limit).photos
+    }
+}
