+++ app/src/main/java/mrcoder/instaexplore/jetpackcompose/data/RetrofitInstance.kt
@@ -1,0 +1,16 @@
+package mrcoder.instaexplore.jetpackcompose.data
+
+import ApiService
+import retrofit2.Retrofit
+import retrofit2.converter.gson.GsonConverterFactory
+
+object RetrofitInstance {
+    private const val BASE_URL = "https://api.slingacademy.com/"
+    val api: ApiService by lazy {
+        Retrofit.Builder()
+            .baseUrl(BASE_URL)
+            .addConverterFactory(GsonConverterFactory.create())
+            .build()
+            .create(ApiService::class.java)
+    }
+}
