+++ app/src/main/java/mrcoder/instaexplore/jetpackcompose/model/Photo.kt
@@ -1,0 +1,12 @@
+package mrcoder.instaexplore.jetpackcompose.model
+
+data class Photo(val url: String) {
+    // تعریف برابری برای جلوگیری از عکس‌های تکراری
+    override fun equals(other: Any?): Boolean {
+        return other is Photo && other.url == this.url
+    }
+
+    override fun hashCode(): Int {
+        return url.hashCode()
+    }
+}
