+++ app/src/main/java/mrcoder/instaexplore/jetpackcompose/model/PhotoResponse.kt
@@ -1,0 +1,5 @@
+package mrcoder.instaexplore.jetpackcompose.model
+
+data class PhotoResponse(val photos: List<Photo>)
+
+
