+++ app/src/main/java/mrcoder/instaexplore/jetpackcompose/ui/components/LoadingScreen.kt
@@ -1,0 +1,30 @@
+package mrcoder.instaexplore.jetpackcompose.ui.components
+
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.size
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+import com.airbnb.lottie.compose.LottieAnimation
+import com.airbnb.lottie.compose.LottieCompositionSpec
+import com.airbnb.lottie.compose.LottieConstants
+import com.airbnb.lottie.compose.rememberLottieComposition
+import mrcoder.instaexplore.jetpackcompose.R
+
+@Composable
+fun LoadingScreen() {
+    val composition by rememberLottieComposition(LottieCompositionSpec.RawRes(R.raw.load))
+
+    Box(modifier = Modifier.fillMaxSize(), contentAlignment = Alignment.Center) {
+        LottieAnimation(
+            composition = composition,
+            modifier = Modifier
+                .size(200.dp) // Adjust the size to your desired value
+                .align(Alignment.Center), // Center it within the Box
+            iterations = LottieConstants.IterateForever // For infinite loop animation
+        )
+    }
+}
