+++ app/src/main/java/mrcoder/instaexplore/jetpackcompose/viewmodel/NetworkStatusViewModelFactory.kt
@@ -1,0 +1,12 @@
+package mrcoder.instaexplore.jetpackcompose.viewmodel
+
+import android.content.Context
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.ViewModelProvider
+
+class NetworkStatusViewModelFactory(private val context: Context) : ViewModelProvider.Factory {
+    override fun <T : ViewModel> create(modelClass: Class<T>): T {
+        @Suppress("UNCHECKED_CAST")
+        return NetworkStatusViewModel(context) as T
+    }
+}
